refactor: move getClient to shared folder
